`refactor(core): reorganize operators/utils modules and extract Config to config.py`
- split monolithic `fxpmath/functions.py` into `fxpmath/operators/` submodules:
  - `core.py`, `arithmetic.py`, `reduction.py`, `array.py`
- convert `fxpmath/functions.py` into a backward-compatible public facade
- replace monolithic `fxpmath/utils.py` with `fxpmath/utils/` package:
  - `common.py`, `parse.py`, `repr.py`, `bitwise.py`, `numeric.py`
  - keep backward-compatible exports via `utils/__init__.py`
- move `Config` out of `objects.py` into new `fxpmath/config.py`
- update imports so `Config` remains available from `fxpmath`, `fxpmath.objects`, and `fxpmath.config`
- preserve public API behavior while reducing module concentration and improving internal structure

Validation:
- full test suite passed: `118 passed, 1 warning`
